+++ src/examples/common/utility.h
@@ -7,7 +7,7 @@
 /**
  * Maps a value from the range [0, 100] to [min, max].
  */
-uint8_t map(uint8_t value, uint32_t min, uint32_t max);
+uint32_t map(uint8_t value, uint32_t min, uint32_t max);
 
 /**
  * Maps a value from the range [min, max] to [0, 100].
